chore: reorgenized code
